requirements.txt
mistralai
flask
Flask-RESTful
dotenv
pinecone
langchain-mistralai
langchain
flask-cors

pypdf
python-docx
